harsha_urlhaus_etl/requirements.txt
pymongo
python-dotenv
requests
cloudscraper
